% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP.R
\name{umapgp}
\alias{umapgp}
\title{UMAPgp package}
\usage{
umapgp(var, data)
}
\arguments{
\item{var}{A variable of interest from the dataset; it should be categorical.}

\item{data}{The dataset to be used; it should be a dataframe with no rownames.}

\item{path}{path to filename}
}
\value{
object of class umap, containing atleast a component with an embedding and a component with configuration settings
}
\description{
Computes Uniform Manifold Approximation and Projection (UMAP) while preserving global structures.
}
\examples{
umapgp(var="Species",data=iris)
}
\references{
Robert E Tarjan and Mihalis Yannakis.(1984). Simple Linear-time algorithms to test chordality of graphs, test acyclicity
of hypergraphs.
}
