% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP.R
\name{umapgp}
\alias{umapgp}
\title{UMAPgp package}
\usage{
umapgp(var, data)
}
\arguments{
\item{var}{A variable of interest from the dataframe, it should be categorical.}

\item{data}{The dataset to be used for analysis, it should be a dataframe.}
}
\value{
object of class umap, containing atleast a component with an embedding and a component with configuration settings
}
\description{
Computes a manifold approximation and projection.
}
\examples{
umapgp(var="Species",data=iris)
}
