% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP.R
\name{umapgp}
\alias{umapgp}
\title{package}
\usage{
umapgp(x, col, data)
}
\arguments{
\item{x}{variable}

\item{col}{variable}

\item{data}{dataframe}
}
\value{
object of class umap, containing atleast a component with an embedding and a component with configuration settings
}
\description{
Computes a manifold approximation and projection
}
\examples{
umapgp(ID,species,data=datasets::iris)

}
