% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPgp.R
\docType{package}
\name{UMAPgp}
\alias{UMAPgp}
\alias{UMAPgp-package}
\title{UMAPgp: Uniform Manifold Approximation and Projection Global Preservation}
\description{
Computes Uniform Manifold Approximation and Projection (UMAP) while preserving global structures.

The UMAPgp package provides one important function:
umapgp
}
\section{UMAPgp functions}{

The UMAPgp function is umapgp
}

\author{
Beatrice, Pujan, Catherine
}
